08/05/2024 09:34 - Correcion de errores

+++ BROLRRHH.Core/Responses/DocumentoResponse/usp_ConsultarDocumento_Response.cs
@@ -11,6 +11,6 @@ namespace BROLRRHH.Core.Responses.DocumentoResponse
         public String? codDoc { get; set; }
         public String? descrCorta { get; set; }
         public String? descrLarga { get; set; }
-        public Int16? longitud { get; set; }
+        public int? longitud { get; set; }
     }
 }
